backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.1.0
SQLAlchemy==2.0.36
alembic==1.13.2
python-multipart==0.0.9
httpx==0.27.2
Jinja2==3.1.4
python-dotenv==1.0.1
loguru==0.7.2
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.0
tqdm==4.66.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
google-generativeai==0.5.4
